+++ source/glest_game/menu/menu_state_custom_game.cpp
@@ -253,9 +253,8 @@ MenuStateCustomGame::MenuStateCustomGame(Program *program, MainMenu *mainMenu,
 	listBoxTileset.init(xoffset+460, mapPos, 150);
     //listBoxTileset.setItems(results);
 	setupTilesetList("");
-    //srand((unsigned int)time(NULL));
 	Chrono seed(true);
-	srand(seed.getCurTicks());
+	srand((unsigned int)seed.getCurTicks());
 
 	listBoxTileset.setSelectedItemIndex(rand() % listBoxTileset.getItemCount());
 
@@ -1568,9 +1567,8 @@ void MenuStateCustomGame::PlayNow(bool saveGame) {
 
 				// Max 1000 tries to get a random, unused faction
 				for(int findRandomFaction = 1; findRandomFaction < 1000; ++findRandomFaction) {
-					//srand((unsigned int)time(NULL) + findRandomFaction);
 					Chrono seed(true);
-					srand(seed.getCurTicks() + findRandomFaction);
+					srand((unsigned int)seed.getCurTicks() + findRandomFaction);
 
 					int selectedFactionIndex = rand() % listBoxFactions[i].getItemCount();
 					string selectedFactionName = listBoxFactions[i].getItem(selectedFactionIndex);
